chore: rename prettier config files to `.cjs`
