+++ packages/lsp0-contracts/.eslintrc.cjs
