+++ packages/lsp1-contracts/.eslintrc.cjs
