+++ packages/lsp2-contracts/.eslintrc.cjs
