+++ packages/lsp3-contracts/.eslintrc.cjs
