+++ packages/lsp5-contracts/constants.ts
