+++ packages/lsp6-contracts/.eslintrc.cjs
