+++ packages/lsp7-contracts/constants.ts
